feature/JENKINS-38339-downstream-job-links-2 * Move actions to pipeline-api-impl
