+++ blueocean-pipeline-api-impl/src/main/java/io/jenkins/blueocean/rest/impl/pipeline/FlowNodeWrapper.java
@@ -3,8 +3,7 @@ package io.jenkins.blueocean.rest.impl.pipeline;
 import hudson.model.Result;
 import io.jenkins.blueocean.rest.model.BlueRun.BlueRunResult;
 import io.jenkins.blueocean.rest.model.BlueRun.BlueRunState;
-import io.jenkins.blueocean.service.embedded.rest.BluePipelineAction;
-import io.jenkins.blueocean.service.embedded.rest.NodeDownstreamBuildAction;
+import io.jenkins.blueocean.listeners.BluePipelineAction;
 import org.jenkinsci.plugins.workflow.actions.ErrorAction;
 import org.jenkinsci.plugins.workflow.graph.AtomNode;
 import org.jenkinsci.plugins.workflow.graph.FlowNode;
